pyproject.toml
[project]
name = "asyn-simulations"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "acpype>=2023.10.27",
    "mdanalysis>=2.9.0",
    "networkx>=3.6.1",
    "numpy>=2.4.4",
    "parmed>=4.3.1",
    "pyyaml>=6.0.3",
    "rdkit>=2025.9.6",
    "seaborn>=0.13.2",
]
